+++ src/prefabs/structures/RichTextInput/options/index.ts
@@ -75,8 +75,14 @@ export const richTextOptions = {
     },
   }),
 
-  label: variable('Label', { value: [''] }),
-  value: variable('Value', { value: [''] }),
+  label: variable('Label', {
+    value: [''],
+    configuration: { allowFormatting: false, allowPropertyName: true },
+  }),
+  value: variable('Value', {
+    value: [''],
+    configuration: { allowFormatting: false },
+  }),
   showBold: toggle('Bold', { value: true }),
   showItalic: toggle('Italic', { value: true }),
   showUnderlined: toggle('Underlined', { value: true }),