+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [2.203.2](https://github.com/bettyblocks/material-ui-component-set/compare/v2.203.1...v2.203.2) (2025-02-06)
+
+
+### Bug Fixes
+
+* disable formatting for value and ensure property name is selectable for label ([85a321b](https://github.com/bettyblocks/material-ui-component-set/commit/85a321b0b42dc25d88187531e75562978eef01fc))
+* ensure property names can be selected ([bd7f37b](https://github.com/bettyblocks/material-ui-component-set/commit/bd7f37b316ff35b67a824522408440b8c5720107))
+
+## [2.203.1](https://github.com/bettyblocks/material-ui-component-set/compare/v2.203.0...v2.203.1) (2025-02-06)
+
+
+### Bug Fixes
+
+* disable formatting for componentoption value and label for inputcomponents ([605d584](https://github.com/bettyblocks/material-ui-component-set/commit/605d584d6010ba4a4e9d0a17e5267c542b43265b))
+* remove afterDeleteHook ([512bc1b](https://github.com/bettyblocks/material-ui-component-set/commit/512bc1b2af1ada4ba73ee03c4cdca38ad9d3559f))
+
 # [2.203.0](https://github.com/bettyblocks/material-ui-component-set/compare/v2.202.0...v2.203.0) (2025-02-04)